chore: Changed where we initialize the RNTP service
- This way, we ensure stores that rely on RNTP to be initialized to hydrate after we can confirm RNTP has been initialized.
